Consolidate all examples into the single example package
Description: Three parallel example-* directories was the wrong repo structure. Fold example-view and example-ssr into the existing src/example package so there is one place for examples: a simple plgg-view gallery, server-side rendering, and client-side rendering (sharing one App). plgg-view's in-package example.tsx stays as-is.

Changes: - Remove src/example-view and src/example-ssr (and their sh scripts)
- src/example now depends on plgg-view + plgg-web and contains: view/ (simple plgg-view components), App.tsx (shared), ssr/server.ts (SSR + serves client bundle), csr/client.tsx (CSR; nullable #root lifted via fromNullable/Option), specs, README; modeling/ (plgg) kept
- example package is now a private demo; vite builds the CSR client bundle (dist/client.js)
- sh: add test-example.sh; npm-install/check-all drop example-view/ssr and run the example check after build.sh

Test Planning: example tsc clean; 7 tests pass (5 plgg-view gallery + SSR/CSR isomorphism); client bundle builds and mounts App in happy-dom; server demo serves the SSR document + /client.js (verified via curl)

Release Preparation: None (experimental POC)
