+++ sh/test-example.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -eu
 REPO_ROOT=$(git rev-parse --show-toplevel) && cd $REPO_ROOT
 
-echo "=== Running 'npm run tsc' in src/example-ssr ==="
-cd $REPO_ROOT/src/example-ssr && npm run tsc
+echo "=== Running 'npm run test' in src/example ==="
+cd $REPO_ROOT/src/example && npm run test
 echo "\n=== All shell scripts have been executed successfully ==="
