--- sh/test-watch-example-view.sh
@@ -1,5 +1,0 @@
-#!/bin/sh -eu
-REPO_ROOT=$(git rev-parse --show-toplevel) && cd $REPO_ROOT
-
-echo "=== Running 'npm run test:watch' in src/example-view ==="
-cd $REPO_ROOT/src/example-view && npm run test:watch
