+++ sh/test-watch-example.sh
@@ -1,0 +1,5 @@
+#!/bin/sh -eu
+REPO_ROOT=$(git rev-parse --show-toplevel) && cd $REPO_ROOT
+
+echo "=== Running 'npm run test:watch' in src/example ==="
+cd $REPO_ROOT/src/example && npm run test:watch
