--- src/example-ssr/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@plgg/example-ssr",
-  "version": "0.0.1",
-  "description": "Isomorphic SSR + CSR demo built with plgg-web and plgg-view",
-  "scripts": {
-    "build": "vite build",
-    "test": "tsc --noEmit && vitest --run",
-    "tsc": "tsc --noEmit",
-    "tsc:watch": "tsc --noEmit --watch",
-    "test:watch": "vitest",
-    "serve": "tsx src/server.ts"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/qmu/plgg.git"
-  },
-  "author": "qmu",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/qmu/plgg/issues"
-  },
-  "homepage": "https://github.com/qmu/plgg#readme",
-  "dependencies": {
-    "plgg": "file:../plgg",
-    "plgg-view": "file:../plgg-view",
-    "plgg-web": "file:../plgg-web"
-  },
-  "devDependencies": {
-    "@types/node": "^25.6.0",
-    "@vitest/coverage-v8": "^4.1.5",
-    "happy-dom": "^15.0.0",
-    "tsx": "^4.0.0",
-    "typescript": "^6.0.3",
-    "vite": "^8.0.10",
-    "vitest": "^4.1.5"
-  }
-}
