--- src/example-ssr/tsconfig.json
@@ -1,39 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2021",
-    "module": "NodeNext",
-    "lib": [
-      "ES2021",
-      "DOM",
-      "DOM.Iterable"
-    ],
-    "types": [
-      "node"
-    ],
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "rootDir": "src",
-    "outDir": "dist",
-    /* JSX resolves to the installed plgg-view package (real consumer). */
-    "jsx": "react-jsx",
-    "jsxImportSource": "plgg-view",
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noUncheckedIndexedAccess": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "allowUnusedLabels": false,
-    "allowUnreachableCode": false,
-    "forceConsistentCasingInFileNames": true,
-    "exactOptionalPropertyTypes": true,
-    "skipLibCheck": true,
-    "isolatedModules": true,
-    "allowJs": false,
-    "erasableSyntaxOnly": true
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
