--- src/example-view/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@plgg/example-view",
-  "version": "0.0.1",
-  "description": "Example client-side app built with the plgg-view library",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.es.js",
-  "exports": {
-    "import": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.es.js"
-    },
-    "require": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.cjs.js"
-    }
-  },
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "vite build",
-    "test": "tsc --noEmit && vitest --run",
-    "tsc": "tsc --noEmit",
-    "tsc:watch": "tsc --noEmit --watch",
-    "test:watch": "vitest",
-    "coverage": "vitest --run --coverage",
-    "coverage:watch": "vitest --coverage"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/qmu/plgg.git"
-  },
-  "author": "qmu",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/qmu/plgg/issues"
-  },
-  "homepage": "https://github.com/qmu/plgg#readme",
-  "dependencies": {
-    "plgg": "file:../plgg",
-    "plgg-view": "file:../plgg-view"
-  },
-  "devDependencies": {
-    "@types/node": "^25.6.0",
-    "@vitest/coverage-v8": "^4.1.5",
-    "typescript": "^6.0.3",
-    "vite": "^8.0.10",
-    "vite-plugin-dts": "^5.0.0",
-    "vitest": "^4.1.5"
-  }
-}
