--- src/example-view/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./TodoItem";
-export * from "./TodoList";
-export * from "./Layout";
