--- src/example-view/src/examples/index.ts
@@ -1,7 +1,0 @@
-// A gallery of small, focused examples — each demonstrates one capability of
-// writing client-side UI with plgg-view.
-export * from "./Greeting";
-export * from "./Badge";
-export * from "./Card";
-export * from "./Menu";
-export * from "./Notice";
