--- src/example-view/src/index.ts
@@ -1,5 +1,0 @@
-// Public surface of the example app. The Node demo (`demo.ts`) is excluded — it
-// has side effects.
-export * from "./components";
-export * from "./examples";
-export * from "./App";
