--- src/example-view/tsconfig.json
@@ -1,43 +1,0 @@
-{
-  "compilerOptions": {
-    /* Compiler Option */
-    "target": "ES2021",
-    "module": "NodeNext",
-    "lib": [
-      "ES2021",
-      "DOM",
-      "DOM.Iterable"
-    ],
-    "types": [
-      "node"
-    ],
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "rootDir": "src",
-    "outDir": "dist",
-    /* JSX resolves to the *installed* plgg-view package (not a path alias): the
-       automatic runtime pulls `jsx`/`jsxs`/`Fragment` and the `JSX` namespace
-       from `plgg-view/jsx-runtime`, exactly as a real consumer would. */
-    "jsx": "react-jsx",
-    "jsxImportSource": "plgg-view",
-    /* Type Checking Rules */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noUncheckedIndexedAccess": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "allowUnusedLabels": false,
-    "allowUnreachableCode": false,
-    "forceConsistentCasingInFileNames": true,
-    "exactOptionalPropertyTypes": true,
-    "skipLibCheck": true,
-    "isolatedModules": true,
-    "allowJs": false,
-    "erasableSyntaxOnly": true
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
