+++ src/example/package.json
@@ -1,20 +1,8 @@
 {
   "name": "@plgg/example",
   "version": "0.0.2",
-  "description": "Example of using plgg",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.es.js",
-  "exports": {
-    "import": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.es.js"
-    },
-    "require": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.cjs.js"
-    }
-  },
-  "types": "dist/index.d.ts",
+  "private": true,
+  "description": "Examples for the plgg monorepo: plgg modeling, plgg-view, and plgg-web SSR/CSR",
   "scripts": {
     "build": "vite build",
     "test": "tsc --noEmit && vitest --run",
@@ -22,7 +10,8 @@
     "tsc:watch": "tsc --noEmit --watch",
     "test:watch": "vitest",
     "coverage": "vitest --run --coverage",
-    "coverage:watch": "vitest --coverage"
+    "coverage:watch": "vitest --coverage",
+    "serve": "tsx src/ssr/server.ts"
   },
   "repository": {
     "type": "git",
@@ -36,14 +25,17 @@
   "homepage": "https://github.com/qmu/plgg#readme",
   "dependencies": {
     "glob": "^13.0.6",
-    "plgg": "file:../plgg"
+    "plgg": "file:../plgg",
+    "plgg-view": "file:../plgg-view",
+    "plgg-web": "file:../plgg-web"
   },
   "devDependencies": {
     "@types/node": "^25.6.0",
     "@vitest/coverage-v8": "^4.1.5",
+    "happy-dom": "^15.0.0",
+    "tsx": "^4.0.0",
     "typescript": "^6.0.3",
     "vite": "^8.0.10",
-    "vite-plugin-dts": "^5.0.0",
     "vitest": "^4.1.5"
   }
 }
