+++ src/example/src/index.ts
@@ -1,1 +1,5 @@
+// Non-side-effect surface (the ssr/server and csr/client entries are excluded —
+// they run, they aren't imported).
 export * from "./modeling";
+export * from "./view";
+export * from "./App";
