+++ src/example/src/view/Badge.tsx
