+++ src/example/src/view/Card.tsx
