+++ src/example/src/view/Notice.tsx
