+++ src/example/src/view/index.ts
@@ -1,0 +1,8 @@
+// A gallery of small, focused plgg-view examples — each demonstrates one
+// capability of authoring views as .tsx components (no SSR/DOM here; the result
+// is the pure-data VNode tree).
+export * from "./Greeting";
+export * from "./Badge";
+export * from "./Card";
+export * from "./Menu";
+export * from "./Notice";
