+++ src/example/tsconfig.json
@@ -3,11 +3,22 @@
     /* Compiler Option */
     "target": "ES2021",
     "module": "NodeNext",
+    "lib": [
+      "ES2021",
+      "DOM",
+      "DOM.Iterable"
+    ],
+    "types": [
+      "node"
+    ],
     "declaration": true,
     "declarationMap": true,
     "sourceMap": true,
     "rootDir": "src",
     "outDir": "dist",
+    /* JSX resolves to the installed plgg-view package (real consumer). */
+    "jsx": "react-jsx",
+    "jsxImportSource": "plgg-view",
     /* Type Checking Rules */
     "strict": true,
     "noUnusedLocals": true,
