+++ src/example/vite.config.ts
@@ -1,34 +1,27 @@
 /// <reference types="vitest" />
 
 import { defineConfig } from "vite";
-import dts from "vite-plugin-dts";
 
 export default defineConfig({
+  // `build` produces the CSR client bundle (dist/client.js) that the SSR server
+  // serves at /client.js. plgg-view and plgg-web/client are bundled in (not
+  // externalized) so it runs in the browser; JSX config is read from tsconfig.
+  build: {
+    outDir: "dist",
+    emptyOutDir: true,
+    lib: {
+      entry: "src/csr/client.tsx",
+      fileName: () => "client.js",
+      formats: ["es"],
+    },
+    rollupOptions: {
+      external: [],
+      output: { exports: "named" },
+    },
+  },
   test: {
     coverage: {
       all: true,
     },
   },
-  build: {
-    outDir: "dist",
-    lib: {
-      entry: "src/index.ts",
-      name: "plgg-example",
-      fileName: (format) => `index.${format}.js`,
-      formats: ["es", "cjs"],
-    },
-    rollupOptions: {
-      external: [],
-      output: {
-        globals: {},
-        exports: "named",
-      },
-    },
-  },
-  plugins: [
-    dts({
-      insertTypesEntry: true,
-      rollupTypes: true,
-    }),
-  ],
 });
